refactor: 코드리뷰 반영

+++ src/main/kotlin/com/nextstep/lotto/domain/LottoResult.kt
@@ -4,6 +4,6 @@ class LottoResult(val result: Map<Rank, Int>) {
     fun calculateRatio(purchaseMoney: Long): Double {
         val totalWinningMoney = result.map { (rank, count) -> count * rank.winningMoney }.sum()
 
-        return (totalWinningMoney / purchaseMoney).toDouble()
+        return (totalWinningMoney.toDouble() / purchaseMoney.toDouble())
     }
 }
